project/broadcaster/requirements.txt
slackclient==1.3.2
asyncio-nats-client
asyncio-nats-streaming
nest_asyncio
